requirements.txt
comet_ml
roboflow
ultralytics
opencv-python
supervision
scikit-learn
numpy
more-itertools
torch
transformers
fastapi
uvicorn
umap-learn